refactor to move spartan link processing into a hook
