+++ react/javascript/demo-embeds/src/hooks/index.js
@@ -27,3 +27,4 @@ export * from './use_all_dashboards'
 export * from './use_all_looks'
 export * from './use_all_explores'
 export * from './use_listen_embed_events'
+export * from './use_spartan_link'
